+++ semialign-indexed/src/Data/Semialign/Indexed.hs
@@ -5,6 +5,8 @@
 -- | Zipping and aligning of indexed functors.
 module Data.Semialign.Indexed (
     SemialignWithIndex (..),
+    -- * Diffing
+    idiff, idiffNoEq
     ) where
 
 import Prelude hiding (zip, zipWith)
@@ -13,6 +15,7 @@ import Control.Lens (FunctorWithIndex (imap))
 
 import Data.Align
 import Data.These
+import Data.Witherable (Filterable (..))
 
 -- Instances
 import Control.Applicative   (ZipList)
@@ -30,6 +33,9 @@ import qualified Data.HashMap.Lazy as HM
 import qualified Data.IntMap       as IntMap
 import qualified Data.Map          as Map
 import qualified Data.Vector       as V
+
+-- $setup
+-- >>> import Data.Map (fromList)
 
 -- | Indexed version of 'Semialign'.
 class (FunctorWithIndex i f, Semialign f) => SemialignWithIndex i f | f -> i where
@@ -96,3 +102,41 @@ instance (Eq k, Hashable k) => SemialignWithIndex k (HashMap k) where
 
 instance SemialignWithIndex Int Vector where
     izipWith = V.izipWith
+
+-------------------------------------------------------------------------------
+-- diffing
+-------------------------------------------------------------------------------
+
+-- | Diff two indexed structures:
+--
+-- >>> idiff [1, 2, 3, 4, 5] [1, 3, 3, 7]
+-- [(1,Just 3),(3,Just 7),(4,Nothing)]
+--
+-- If your structure is explicitly indexed ('Map'-like), the result
+-- type is awkward, and you may prefer 'Data.Semialign.diff' instead:
+--
+-- >>> idiff (fromList [(0, 1), (2, 3)]) (fromList [(0, 0), (1, 2)])
+-- fromList [(0,(0,Just 0)),(1,(1,Just 2)),(2,(2,Nothing))]
+idiff
+  :: (SemialignWithIndex i f, Filterable f, Eq a)
+  => f a -> f a -> f (i, Maybe a)
+idiff = (catMaybes .) . ialignWith merge where
+  merge i (This _) = Just (i, Nothing)
+  merge i (That new) = Just (i, Just new)
+  merge i (These old new)
+    | old == new = Nothing
+    | otherwise = Just (i, Just new)
+
+-- | Diff two indexed structures without requiring an 'Eq'
+-- instance. Instead, assume a new value wherever the structures
+-- align:
+--
+-- >>> (fmap . fmap) ($ 3) (idiffNoEq [(+1)] [(*2), (const 0)] !! 1)
+-- (1,Just 0)
+idiffNoEq
+  :: (SemialignWithIndex i f, Filterable f)
+  => f a -> f a -> f (i, Maybe a)
+idiffNoEq = (catMaybes .) . ialignWith merge where
+  merge i (This _) = Just (i, Nothing)
+  merge i (That new) = Just (i, Just new)
+  merge i (These _ new) = Just (i, Just new)