+++ semialign-indexed/CHANGELOG.md
@@ -1,3 +1,8 @@
+# Unreleased
+
+* Add `Data.Semialign.Indexed.diff`.
+* Add `Data.Semialign.Indexed.diffNoEq`.
+
 # 1
 
 Split out of `these` package.