+++ semialign/CHANGELOG.md
@@ -1,3 +1,9 @@
+# Unreleased
+
+* Add `Data.Semialign.diff`.
+* Add `Data.Semialign.diffNoEq`.
+* Add `Data.Semialign.patch`.
+
 # 1
 
 Split out of `these` package.